Log welcome message on log level LIFECYCLE
Logging the welcome message to System.out without going through the
logging system broke several users CI builds because they expected
nothing to be logged when passing the quiet option. This changes the
logging of the welcome message to use the logging system on log level LIFECYCLE
which causes the welcome message not to be logged anymore when using the quiet
option.

Issue: #5213